chore: update tests and strauss
