+++ composer.json
@@ -38,9 +38,6 @@
 			"namespace_prefix": "WPGraphQL\\Login\\Vendor\\",
 			"classmap_prefix": "GraphQL_Login_",
 			"constant_prefix": "GRAPHQL_LOGIN_",
-			"delete_vendor_files": false,
-			"delete_vendor_packages": false,
-			"update_call_sites": false,
 			"include_modified_date": false,
 			"packages": [
 				"axepress/axewp-common",
@@ -55,8 +52,7 @@
 			"exclude_from_prefix": {
 				"namespaces": [
 					"WPGraphQL\\"
-				],
-				"file_patterns": []
+				]
 			}
 		}
 	},
@@ -75,6 +71,9 @@
 		"psr-4": {
 			"Tests\\WPGraphQL\\Login\\": "tests/phpunit/"
 		}
+	},
+	"replace": {
+		"php-extended/polyfill-php80-str-utils": "*"
 	},
 	"require": {
 		"php": ">=8.2",
@@ -107,7 +106,7 @@
 		"phpstan": "vendor/bin/phpstan analyze --ansi --memory-limit=1G",
 		"test": "vendor/bin/phpunit -c phpunit.xml.dist",
 		"strauss-install": [
-			"test -f ./bin/strauss.phar || curl -o bin/strauss.phar -L -C - https://github.com/BrianHenryIE/strauss/releases/download/0.26.3/strauss.phar"
+			"test -f ./bin/strauss.phar || curl -o bin/strauss.phar -L -C - https://github.com/BrianHenryIE/strauss/releases/download/0.28.1/strauss.phar"
 		],
 		"prefix-namespaces": [
 			"@strauss-install",
