+++ tests/phpunit/Integration/Mutations/ProviderMutationsGoogleTest.php
@@ -5,50 +5,12 @@
  * @package Tests\WPGraphQL\Login\Integration\Mutations
  */
 
-use Mockery as m;
-use WPGraphQL\Login\Auth\ProviderConfig\OAuth2\Google as OAuth2Google;
-use WPGraphQL\Login\Auth\ProviderConfig\OAuth2\OAuth2Config;
+declare( strict_types = 1 );
+
+namespace Tests\WPGraphQL\Login\Integration\Mutations;
+
+use Tests\WPGraphQL\Login\Fixtures\FooGoogleProviderConfig;
 use WPGraphQL\Login\Auth\User;
-use WPGraphQL\Login\Vendor\League\OAuth2\Client\Provider\Google;
-
-/**
- * A Google provider with mocked resource owner details.
- */
-class FooGoogleProvider extends Google {
-	/**
-	 * {@inheritDoc}
-	 */
-	protected function fetchResourceOwnerDetails( $token ) {
-		return json_decode( '{"sub": 12345, "name": "mock_name", "given_name": "mock_first_name", "family_name": "mock_last_name", "email": "mock_email@mockdomain.com", "email_verified": true, "picture": "mock_image_url", "hd":"mockdomain.com"}', true );
-	}
-}
-
-/**
- * A Google provider config wired to the mocked provider and HTTP client.
- */
-class FooGoogleProviderConfig extends OAuth2Google {
-	/**
-	 * {@inheritDoc}
-	 */
-	public function __construct() {
-		OAuth2Config::__construct( FooGoogleProvider::class );
-
-		// Mock and set the http client on the provider.
-		$response = m::mock( 'WPGraphQL\Login\Vendor\Psr\Http\Message\ResponseInterface' );
-		$response->shouldReceive( 'getHeader' )
-			->times( 1 )
-			->andReturn( [ 'Content-Type' => 'application/json' ] );
-		$response->shouldReceive( 'getBody' )
-			->times( 1 )
-			->andReturn(
-				\WPGraphQL\Login\Vendor\GuzzleHttp\Psr7\Utils::streamFor( '{"access_token":"mock_access_token","token_type":"bearer","expires_in":3600}' )
-			);
-
-		$http_client = m::mock( 'WPGraphQL\Login\Vendor\GuzzleHttp\ClientInterface' );
-		$http_client->shouldReceive( 'send' )->times( 1 )->andReturn( $response );
-		$this->provider->setHttpClient( $http_client );
-	}
-}
 
 /**
  * Tests the Google provider mutations.
